move load script to scripts js
