+++ blocks/form/form.js
@@ -1,5 +1,5 @@
 import { fetchPlaceholders, readBlockConfig } from '../../scripts/lib-franklin.js';
-import { createElement } from '../../scripts/scripts.js';
+import { createElement, loadScript } from '../../scripts/scripts.js';
 
 function constructPayload(form) {
   const payload = {};
@@ -248,26 +248,6 @@ async function createForm(formURL) {
     form.append(formField);
   });
   return form;
-}
-
-/**
- * load a script by adding to page head
- * @param {string} url the script src url
- * @param {string} type the script type
- * @param {function} callback a funciton to callback after loading
- */
-export function loadScript(url, type, callback) {
-  const head = document.querySelector('head');
-  let script = head.querySelector(`script[src="${url}"]`);
-  if (!script) {
-    script = document.createElement('script');
-    script.src = url;
-    if (type) script.setAttribute('type', type);
-    head.append(script);
-    script.onload = callback;
-    return script;
-  }
-  return script;
 }
 
 export default async function decorate(block) {
