+++ scripts/scripts.js
@@ -56,6 +56,26 @@ export function createElement(tagName, classes, props, html) {
   return elem;
 }
 
+/**
+ * load a script by adding to page head
+ * @param {string} url the script src url
+ * @param {string} type the script type
+ * @param {function} callback a funciton to callback after loading
+ */
+export function loadScript(url, type, callback) {
+  const head = document.querySelector('head');
+  let script = head.querySelector(`script[src="${url}"]`);
+  if (!script) {
+    script = document.createElement('script');
+    script.src = url;
+    if (type) script.setAttribute('type', type);
+    head.append(script);
+    script.onload = callback;
+    return script;
+  }
+  return script;
+}
+
 function buildNewsColumns(main) {
   if (!document.body.classList.contains('news')) {
     return;
